Add files via upload
Select one of the following two problems:

I've selected the FDA problem
1- Implement a classifier based on FDA and compare it with the Baysian-Gaussian classifier you
developed in Assignment 1-b). Show the effect of multiple input features on the final accuracy of
both classifiers.

Because the FDA's main goal is reduce multiple features to that of a linear classifier, we will only compare the FDA to the Baysian Gaussian Classifier. We will do so by comparing the accuracies as well as ROC curves of the best features from both the Baysian and the Gaussian Classifier for multiple features. 

The following assumption is made with the Gaussian classifier, when increasing the number of features, the amount of time taken to train and test the classifier increases drastically. This is due to the g-matrix used for boundary calculation increasing from  1 dimensions to 2, 3, and 4 dimensions. Thus we will only compare the FDA with the best classifier from Gaussian using 2 features, not more, as both the accuracy and the speed of the classfier as high.

In the FDA, most classifiers output 100% accuracy as well as a perfect ROC curve due to the  maximum seperation of data. Using the same logic for FDA, although the data ultimately collapses to one dimension calculating the projection of 2 features onto one dimension values is slightly faster than 3 and 4 features. Thus we'll select our best classifier from the projections of the two features. 

I selected the FDA classifier which combines feature 2 and 3 since the seperation between the two points is vast as shown by the gaussian, g-values, and g-decision bounday curves before.